chore: add tests for spcp.oidc.service.cp
